# chat-backend/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "TAI后端服务",
    "title": "TAI Backend API",
    "version": "1.0.0"
  },
  "host": "localhost:9090",
  "basePath": "/api/v1",
  "paths": {
    "/chat/conversations": {
      "get": {
        "description": "分页获取用户的对话列表，支持页码和每页数量参数",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "获取对话列表",
        "operationId": "getConversations",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ConversationListResponse",
            "schema": {
              "$ref": "#/definitions/ConversationListResponse"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "post": {
        "description": "根据配置创建一个新的对话",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "创建对话",
        "operationId": "createConversation",
        "parameters": [
          {
            "description": "对话设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "对话设置",
              "type": "object",
              "$ref": "#/definitions/ConversationSettings"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation",
            "schema": {
              "$ref": "#/definitions/Conversation"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/conversations/{id}": {
      "delete": {
        "description": "根据对话ID删除指定的对话记录和相关数据",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "删除对话",
        "operationId": "deleteConversation",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/conversations/{id}/history": {
      "get": {
        "description": "根据对话ID获取该对话的完整消息历史记录",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "获取对话历史",
        "operationId": "getConversationHistory",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ConversationHistoryResponse",
            "schema": {
              "$ref": "#/definitions/ConversationHistoryResponse"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/conversations/{id}/settings": {
      "get": {
        "description": "根据对话ID获取该对话的详细设置信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "获取对话设置",
        "operationId": "getConversationSettings",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ConversationSettings",
            "schema": {
              "$ref": "#/definitions/ConversationSettings"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "put": {
        "description": "根据对话ID更新该对话的设置信息，如模型配置、参数等",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "更新对话设置",
        "operationId": "updateConversationSettings",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "对话设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "对话设置",
              "type": "object",
              "$ref": "#/definitions/ConversationSettings"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageWithDataResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/messages": {
      "post": {
        "description": "向指定对话发送消息并以SSE流式方式返回AI响应",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "发送消息",
        "operationId": "sendMessage",
        "parameters": [
          {
            "description": "聊天请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "聊天请求",
              "type": "object",
              "$ref": "#/definitions/ChatRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ChatRequest": {
      "description": "ChatRequest 聊天请求",
      "type": "object",
      "required": [
        "conversationId",
        "content",
        "requestId"
      ],
      "properties": {
        "content": {
          "description": "消息内容",
          "type": "string",
          "x-go-name": "Content"
        },
        "conversationId": {
          "description": "会话ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ConversationID"
        },
        "requestId": {
          "description": "请求唯一uuid",
          "type": "string",
          "x-go-name": "RequestID"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "Conversation": {
      "description": "Conversation 对话信息",
      "type": "object",
      "required": [
        "name",
        "desc",
        "model_id",
        "temperature",
        "top_p",
        "presence_penalty",
        "frequency_penalty",
        "responseType",
        "stream",
        "knowledge_base_ids",
        "contextLimit",
        "id"
      ],
      "properties": {
        "contextLimit": {
          "description": "上下文限制（消息数量）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ContextLimit"
        },
        "desc": {
          "description": "对话描述",
          "type": "string",
          "x-go-name": "Desc"
        },
        "frequency_penalty": {
          "description": "重复控制",
          "type": "number",
          "format": "double",
          "x-go-name": "FrequencyPenalty"
        },
        "id": {
          "description": "对话ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "knowledge_base_ids": {
          "description": "知识库ID列表",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "KnowledgeBaseIDs"
        },
        "model_id": {
          "description": "模型ID（数字）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ModelID"
        },
        "name": {
          "description": "对话名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "presence_penalty": {
          "description": "词汇控制",
          "type": "number",
          "format": "double",
          "x-go-name": "PresencePenalty"
        },
        "responseType": {
          "description": "响应类型  text/json",
          "type": "string",
          "x-go-name": "ResponseType"
        },
        "stream": {
          "description": "对话输出模式 true=流式输出, false=非流式输出",
          "type": "boolean",
          "x-go-name": "Stream"
        },
        "temperature": {
          "description": "多样性",
          "type": "number",
          "format": "double",
          "x-go-name": "Temperature"
        },
        "top_p": {
          "description": "采样范围",
          "type": "number",
          "format": "double",
          "x-go-name": "TopP"
        }
      },
      "x-go-package": "chat-backend/models"
    },
    "ConversationHistoryResponse": {
      "description": "ConversationHistoryResponse 对话历史响应",
      "type": "object",
      "required": [
        "conversation_id",
        "messages",
        "total"
      ],
      "properties": {
        "conversation_id": {
          "description": "对话ID",
          "type": "string",
          "x-go-name": "ConversationID"
        },
        "messages": {
          "description": "消息列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MessageRecord"
          },
          "x-go-name": "Messages"
        },
        "total": {
          "description": "消息总数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ConversationListResponse": {
      "description": "ConversationListResponse 对话列表响应",
      "type": "object",
      "required": [
        "conversations",
        "total",
        "page",
        "page_size"
      ],
      "properties": {
        "conversations": {
          "description": "对话列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Conversation"
          },
          "x-go-name": "Conversations"
        },
        "page": {
          "description": "当前页码",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Page"
        },
        "page_size": {
          "description": "每页数量",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PageSize"
        },
        "total": {
          "description": "总数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "chat-backend/models"
    },
    "ConversationSettings": {
      "description": "ConversationSettings 对话设置",
      "type": "object",
      "required": [
        "name",
        "desc",
        "model_id",
        "temperature",
        "top_p",
        "presence_penalty",
        "frequency_penalty",
        "responseType",
        "stream",
        "knowledge_base_ids",
        "contextLimit"
      ],
      "properties": {
        "contextLimit": {
          "description": "上下文限制（消息数量）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ContextLimit"
        },
        "desc": {
          "description": "对话描述",
          "type": "string",
          "x-go-name": "Desc"
        },
        "frequency_penalty": {
          "description": "重复控制",
          "type": "number",
          "format": "double",
          "x-go-name": "FrequencyPenalty"
        },
        "knowledge_base_ids": {
          "description": "知识库ID列表",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "KnowledgeBaseIDs"
        },
        "model_id": {
          "description": "模型ID（数字）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ModelID"
        },
        "name": {
          "description": "对话名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "presence_penalty": {
          "description": "词汇控制",
          "type": "number",
          "format": "double",
          "x-go-name": "PresencePenalty"
        },
        "responseType": {
          "description": "响应类型  text/json",
          "type": "string",
          "x-go-name": "ResponseType"
        },
        "stream": {
          "description": "对话输出模式 true=流式输出, false=非流式输出",
          "type": "boolean",
          "x-go-name": "Stream"
        },
        "temperature": {
          "description": "多样性",
          "type": "number",
          "format": "double",
          "x-go-name": "Temperature"
        },
        "top_p": {
          "description": "采样范围",
          "type": "number",
          "format": "double",
          "x-go-name": "TopP"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "MessageRecord": {
      "description": "MessageRecord 对话消息记录",
      "type": "object",
      "required": [
        "id",
        "role",
        "content",
        "created_at"
      ],
      "properties": {
        "content": {
          "description": "消息内容",
          "type": "string",
          "x-go-name": "Content"
        },
        "created_at": {
          "description": "创建时间",
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "id": {
          "description": "消息ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "role": {
          "description": "角色: user/assistant",
          "type": "string",
          "x-go-name": "Role"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ResponseBody": {
      "description": "ResponseBody 统一响应体结构",
      "type": "object",
      "required": [
        "is_login",
        "code",
        "msg",
        "data",
        "time",
        "code_num",
        "code_msg"
      ],
      "properties": {
        "code": {
          "description": "请求是否成功",
          "type": "boolean",
          "x-go-name": "Code"
        },
        "code_msg": {
          "description": "错误代码消息",
          "type": "string",
          "x-go-name": "CodeMsg"
        },
        "code_num": {
          "description": "数字错误代码",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CodeNum"
        },
        "data": {
          "description": "响应数据",
          "x-go-name": "Data"
        },
        "is_login": {
          "description": "登录状态",
          "type": "boolean",
          "x-go-name": "IsLogin"
        },
        "msg": {
          "description": "响应消息",
          "type": "string",
          "x-go-name": "Message"
        },
        "time": {
          "description": "持续时间（秒）",
          "type": "number",
          "format": "double",
          "x-go-name": "Time"
        }
      },
      "x-go-package": "chat-backend/utils"
    },
...
  },
  "responses": {
...
    "MessageOnlyResponse": {
      "description": "MessageOnlyResponse 仅消息成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "MessageWithDataResponse": {
      "description": "MessageWithDataResponse 消息和数据成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "description": "响应数据",
            "x-go-name": "Data"
          },
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
  }
}